pyproject.toml
[tool.poetry]
name = "sunva-ai"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.dependencies]
python = ">=3.10,<3.12"  # Assuming Python 3.10, adjust as needed

fastapi = "0.104.1"
numpy = "^1.22.0"
groq = "0.9.0"
uvicorn = "0.24.0.post1"
faster-whisper = "1.0.0"
nest-asyncio = "^1.5.8"
websocket-client = "1.6.4"
websockets = "11.0.3"
python-dotenv = "^1.0.0"
TTS = "0.22.0"
pydantic = "^2.8.2"
instructor = "1.3.7"
redis = "5.0.8"
sqlalchemy = "2.0.32"
sqlmodel = "0.0.21"
aiosqlite = '0.20.0'
anthropic = "0.34.1"
openai = "1.41.1"
librosa = "0.10.2.post1"
